Rename scan_path_utils.py to utils.py
